+++ CHANGES.rst
@@ -18,7 +18,7 @@ in progress
 - Add support for XLSX and ODS input formats
 - Add test case for basic nested JSON document
 - Add ``frictionless`` backend
-- Add support for Google Sheets input format
+- Add support for Google Sheets and Parquet input formats
 
 
 2023-01-xx 0.0.0